Fix org-starter-def macro
Formerly, org-starter-def macro didn't work on a file, because
file-exists-p was tested on a file name possibly without a directory.
This has been fixed by using 'if' rather than 'cond' and skipping
the file-exists-p test.

Also, the returned argument was not correct, as the given file
name was returned. Fixed so that the result of
org-starter-define-file/directory is returned.

Also tidied by using ,@ backquoting rather than using apply.