Return a description of the operation from -def-capture macro
Since this function is likely to be evaluated at the time of
experimenting with configuration, it is probably annoying for the user
to get the whole value of org-capture-templates. Instead, I have
changed the macro to return the brief description of the operation,
i.e. whether it has overridden an existing template of the same key or
inserted a new entry. This will be useful for preventing conflicts.