Add descriptive arguments to the agenda function
Add to org-starter-add-agenda-custom-command
the concrete definition of arguments as defined in
org-agenda-custom-commands,
so the user can see them via eldoc.

The arguments are different when a block agenda is being defined,
so I also added org-starter-add-block-agenda-command for such situations.
This is just a wrapper for org-starter-add-agenda-custom-command
with a different set of arguments.

WIP: Also verify the arguments when the user tries to define an agenda
command after startup. This can be disabled by setting
org-starter-verify-agenda-configuration to nil.